refactor: Changed directory structure
