+++ packages/xhear/src/register.mjs
