+++ packages/xhear/tests/stanz-test.spec.js
@@ -1,7 +1,7 @@
 const { test, expect } = require("@playwright/test");
 
 test("stanz", async ({ page }) => {
-  await page.goto("http://localhost:3398/e2e/statics/stanz-test.html");
+  await page.goto("http://localhost:3398/packages/xhear/tests/statics/stanz-test.html");
 
   await page.waitForSelector(".jasmine-suite-detail.jasmine-passed");
 
@@ -14,7 +14,7 @@ test("stanz", async ({ page }) => {
 
 test("xhear sub object", async ({ page }) => {
   await page.goto(
-    "http://localhost:3398/e2e/statics/xhear-sub-object-test.html"
+    "http://localhost:3398/packages/xhear/tests/statics/xhear-sub-object-test.html"
   );
 
   await page.waitForSelector(".jasmine-suite-detail.jasmine-passed");
