+++ packages/xhear/tests/statics/base-methods-test.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>base test</title>
-    <script src="../../dist/xhear.js"></script>
+    <script src="../../../../dist/xhear.js"></script>
   </head>
   <body>
     <div id="target1" data-testid="target1">I am target</div>
