+++ packages/xhear/tests/statics/event-test.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>xhear sub object test</title>
-    <script src="../../dist/xhear.js"></script>
+    <script src="../../../../dist/xhear.js"></script>
     <!-- <script type="module">
       import $ from "../../packages/main/base.mjs";
