+++ packages/xhear/tests/statics/render-test.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
-    <script src="../../dist/xhear.js"></script>
+    <script src="../../../../dist/xhear.js"></script>
     <style>
       .sub-item {
         margin: 10px;
