+++ packages/xhear/tests/statics/stanz-test.html
@@ -7,13 +7,13 @@
     <title>$.stanz test</title>
     <link
       rel="stylesheet"
-      href="../../node_modules/jasmine-core/lib/jasmine-core/jasmine.css"
+      href="../../../../node_modules/jasmine-core/lib/jasmine-core/jasmine.css"
     />
-    <script src="../../node_modules/jasmine-core/lib/jasmine-core/jasmine.js"></script>
-    <script src="../../node_modules/jasmine-core/lib/jasmine-core/jasmine-html.js"></script>
-    <script src="../../node_modules/jasmine-core/lib/jasmine-core/boot0.js"></script>
-    <script src="../../node_modules/jasmine-core/lib/jasmine-core/boot1.js"></script>
-    <script src="../../dist/xhear.js"></script>
+    <script src="../../../../node_modules/jasmine-core/lib/jasmine-core/jasmine.js"></script>
+    <script src="../../../../node_modules/jasmine-core/lib/jasmine-core/jasmine-html.js"></script>
+    <script src="../../../../node_modules/jasmine-core/lib/jasmine-core/boot0.js"></script>
+    <script src="../../../../node_modules/jasmine-core/lib/jasmine-core/boot1.js"></script>
+    <script src="../../../../dist/xhear.js"></script>
   </head>
   <body>
     <script>
@@ -28,19 +28,19 @@
       const test = it;
     </script>
     <script
-      src="../../packages/stanz/test/accessor.test.js"
+      src="../../../stanz/test/accessor.test.js"
       type="module"
     ></script>
     <script
-      src="../../packages/stanz/test/methods.test.js"
+      src="../../../stanz/test/methods.test.js"
       type="module"
     ></script>
     <script
-      src="../../packages/stanz/test/conversion.test.js"
+      src="../../../stanz/test/conversion.test.js"
       type="module"
     ></script>
     <script
-      src="../../packages/stanz/test/watch.test.js"
+      src="../../../stanz/test/watch.test.js"
       type="module"
     ></script>
   </body>
