+++ packages/xhear/tests/statics/watch-test.html
@@ -7,13 +7,13 @@
     <title>xhear sub object test</title>
     <link
       rel="stylesheet"
-      href="../../node_modules/jasmine-core/lib/jasmine-core/jasmine.css"
+      href="../../../../node_modules/jasmine-core/lib/jasmine-core/jasmine.css"
     />
-    <script src="../../node_modules/jasmine-core/lib/jasmine-core/jasmine.js"></script>
-    <script src="../../node_modules/jasmine-core/lib/jasmine-core/jasmine-html.js"></script>
-    <script src="../../node_modules/jasmine-core/lib/jasmine-core/boot0.js"></script>
-    <script src="../../node_modules/jasmine-core/lib/jasmine-core/boot1.js"></script>
-    <!-- <script src="../../dist/xhear.js"></script> -->
+    <script src="../../../../node_modules/jasmine-core/lib/jasmine-core/jasmine.js"></script>
+    <script src="../../../../node_modules/jasmine-core/lib/jasmine-core/jasmine-html.js"></script>
+    <script src="../../../../node_modules/jasmine-core/lib/jasmine-core/boot0.js"></script>
+    <script src="../../../../node_modules/jasmine-core/lib/jasmine-core/boot1.js"></script>
+    <script src="../../../../dist/xhear.js"></script>
     <script>
       const test = it;
     </script>
@@ -27,7 +27,7 @@
       <div id="target2" data-testid="target2">I am target</div>
     </div>
     <script type="module">
-      import $ from "../../../packages/main/base.mjs";
+      // import $ from "../../../packages/main/base.mjs";
 
       describe("test xhear watch", () => {
         test("test watch update", () => {
