+++ packages/xhear/tests/watch-test.spec.js
@@ -1,7 +1,7 @@
 const { test, expect } = require("@playwright/test");
 
 test("xhear sub object", async ({ page }) => {
-  await page.goto("http://localhost:3398/e2e/statics/watch-test.html");
+  await page.goto("http://localhost:3398/packages/xhear/tests/statics/watch-test.html");
 
   await page.waitForSelector(".jasmine-suite-detail.jasmine-passed");
